fix manifests
