+++ src/packages/settings/data-types/repository/data-type-repository-base.ts
@@ -1,6 +1,6 @@
 import { UMB_DATA_TYPE_ITEM_STORE_CONTEXT_TOKEN, UmbDataTypeItemStore } from './item/data-type-item.store.js';
 import { UMB_DATA_TYPE_TREE_STORE_CONTEXT_TOKEN, UmbDataTypeTreeStore } from '../tree/data-type.tree.store.js';
-import { UMB_DATA_TYPE_STORE_CONTEXT_TOKEN, UmbDataTypeStore } from './detail/data-type-detail.store.js';
+import { UMB_DATA_TYPE_STORE_CONTEXT_TOKEN, UmbDataTypeDetailStore } from './detail/data-type-detail.store.js';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { UMB_NOTIFICATION_CONTEXT_TOKEN, UmbNotificationContext } from '@umbraco-cms/backoffice/notification';
 import { UmbRepositoryBase } from '@umbraco-cms/backoffice/repository';
@@ -8,7 +8,7 @@ import { UmbRepositoryBase } from '@umbraco-cms/backoffice/repository';
 export class UmbDataTypeRepositoryBase extends UmbRepositoryBase {
 	protected _init: Promise<unknown>;
 
-	protected _detailStore?: UmbDataTypeStore;
+	protected _detailStore?: UmbDataTypeDetailStore;
 	protected _treeStore?: UmbDataTypeTreeStore;
 	protected _itemStore?: UmbDataTypeItemStore;
